+++ django_vox/__init__.py
@@ -1,3 +1,3 @@
 default_app_config = 'django_vox.apps.VoxConfig'
 
-VERSION = '3.4.0'
+VERSION = '3.5.0'